requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.3
fake-useragent==1.1.3
idna==3.4
numpy==1.25.0
pandas==2.0.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
useragent==0.1.1
Pillow==10.0.0
gevent==22.10.2
gunicorn==20.1.0
# postgresql==12